requirement.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
botocore==1.19.53
chardet==4.0.0
click==7.1.2
Flask==1.1.2
git-remote-codecommit==1.15.1
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
multidict==5.1.0
pyee==7.0.4
python-dateutil==2.8.1
python-dotenv==0.15.0
six==1.15.0
slackclient==2.9.3
slackeventsapi==2.2.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.6.3
